requirements.txt
django==4.2
# django-extensions==3.2.1
# django-model-utils==4.3.1

# Storage
# redis==4.5.5
# django-redis==5.2.0
psycopg2-binary==2.9.5

# django-libsass==0.9
# time-machine==2.9.0
# python-dateutil==2.8.2
djangorestframework==3.14.0

# django-thumbnails==0.7.0
django-positions==0.6.0
# shortuuid==1.0.11
